NP-PRA: Cleanup for PRA
- Also added an extra test for HeaderCarrierExtensions to bring up coverage to 100%
