+++ app/uk/gov/hmrc/uknwauthcheckerapi/utils/RequestBuilderExtensions.scala
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2024 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.uknwauthcheckerapi.utils
+
+import play.api.http.Status.OK
+import play.api.libs.json.Reads
+import uk.gov.hmrc.http.HttpReads.Implicits.*
+import uk.gov.hmrc.http.HttpResponse
+import uk.gov.hmrc.http.client.RequestBuilder
+
+import scala.concurrent.{ExecutionContext, Future}
+
+trait RequestBuilderExtensions extends HttpResponseExtensions {
+  
+  implicit class RequestBuilderExtension(requestBuilder: RequestBuilder) {
+    def executeAndDeserialise[T](implicit ec: ExecutionContext, reads: Reads[T]): Future[T] =
+      requestBuilder
+        .execute[HttpResponse]
+        .flatMap { response =>
+          response.status match {
+            case OK => response.as[T]
+            case _ =>
+              response.error
+          }
+        }
+  }
+}
