+++ test/uk/gov/hmrc/uknwauthcheckerapi/utils/HeaderCarrierExtensionsSpec.scala
@@ -1,0 +1,50 @@
+/*
+ * Copyright 2024 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.uknwauthcheckerapi.utils
+
+import uk.gov.hmrc.http.HeaderCarrier
+import uk.gov.hmrc.uknwauthcheckerapi.controllers.BaseSpec
+import uk.gov.hmrc.uknwauthcheckerapi.generators.TestRegexes
+import uk.gov.hmrc.uknwauthcheckerapi.models.constants.CustomHeaderNames
+
+import java.util.UUID
+
+class HeaderCarrierExtensionsSpec extends BaseSpec with HeaderCarrierExtensions {
+
+  "generateCorrelationId" should {
+    "reuse the existing X-Correlation-Id header when present in HeaderCarrier" in {
+      val uuid = UUID.randomUUID()
+
+      implicit val headerCarrier: HeaderCarrier = new HeaderCarrier(
+        otherHeaders = Seq((CustomHeaderNames.xCorrelationId, uuid.toString))
+      )
+
+      val correlationId: String = generateCorrelationId()
+
+      correlationId shouldBe uuid.toString
+    }
+
+    "generate new X-Correlation-Id header when not present in HeaderCarrier" in {
+
+      implicit val headerCarrier: HeaderCarrier = new HeaderCarrier()
+
+      val correlationId: String = generateCorrelationId()
+
+      correlationId should fullyMatch regex TestRegexes.uuidPattern
+    }
+  }
+}
